Adding updateAccuracyResult() function to update accuracyResult var whenever globe is matching.

+++ SimpleGlobe/Model/Study/PositionTaskCodable.swift
@@ -8,9 +8,9 @@
 
 struct PositionTaskCodable: Encodable {
     var actions: [StudyAction]
-    var accuracyResult: Int
+    var accuracyResult: Float
     
-    init(actions: [StudyAction], accuracyResult: Int) {
+    init(actions: [StudyAction], accuracyResult: Float) {
         self.actions = actions
         self.accuracyResult = accuracyResult
     }
